+++ src/commands/undo.rs
@@ -26,7 +26,7 @@ use crate::core::graph::{make_graph, BranchOids, HeadOid, MainBranchOid};
 use crate::core::mergebase::MergeBaseDb;
 use crate::core::metadata::{
     render_commit_metadata, BranchesProvider, CommitMessageProvider, CommitOidProvider,
-    DifferentialRevisionProvider, RelativeTimeProvider,
+    DifferentialRevisionProvider, HiddenExplanationProvider, RelativeTimeProvider,
 };
 use crate::util::{get_db_conn, get_repo, run_git, GitExecutable};
 
@@ -98,6 +98,7 @@ fn render_cursor_smartlog(
         &[
             &CommitOidProvider::new(true)?,
             &RelativeTimeProvider::new(&repo, SystemTime::now())?,
+            &HiddenExplanationProvider::new(&graph, &event_replayer, event_cursor)?,
             &BranchesProvider::new(&repo, &branch_oid_to_names)?,
             &DifferentialRevisionProvider::new(&repo)?,
             &CommitMessageProvider::new()?,